+++ src/wasmJsMain/npm/hayagriva-wasm/hayagriva_wasm_bg.wasm.d.ts
@@ -1,0 +1,14 @@
+/* tslint:disable */
+/* eslint-disable */
+export const memory: WebAssembly.Memory;
+export const __wbg_bibliographer_free: (a: number, b: number) => void;
+export const bibliographer_bibliography: (a: number) => [number, number];
+export const bibliographer_citation: (a: number, b: number, c: number) => [number, number, number, number];
+export const bibliographer_citation_keys: (a: number) => [number, number];
+export const bibliographer_new: (a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number) => [number, number, number];
+export const __wbindgen_externrefs: WebAssembly.Table;
+export const __wbindgen_free: (a: number, b: number, c: number) => void;
+export const __wbindgen_malloc: (a: number, b: number) => number;
+export const __wbindgen_realloc: (a: number, b: number, c: number, d: number) => number;
+export const __externref_table_dealloc: (a: number) => void;
+export const __wbindgen_start: () => void;